added VRAM for NMI safety and feature compat

+++ demo/levels.c
@@ -32,7 +32,7 @@ uint8_t GetNextMetaTile() {
     return tile;
 }
 
-uint8_t GetNextWrite(const uint8_t step) {
+uint8_t GetNextWrite(const uint16_t step) {
     if (~step & 1) {
         if (step == 0) {
             attr_column++;
@@ -55,6 +55,6 @@ uint8_t GetNextWrite(const uint8_t step) {
     return Metatiles[MetatileBuffer[step >> 1] << 2 | (step & 1)];
 }
 
-uint8_t GetCurrentWrite(const uint8_t step) {
+uint8_t GetCurrentWrite(const uint16_t step) {
     return Metatiles[MetatileBuffer[step >> 1] << 2 | 2 | step & 1];
 }
